requirements.txt
Flask
langchain
python-dotenv
openai
faiss-cpu
requests
beautifulsoup4
